refactor: remove unused accordion component, sorry dom
