--- ui/src/components/Accordion-new/index.js
@@ -1,4 +1,0 @@
-import Accordion from './Accordion';
-import AccordionSection, {AccordionSectionHeader} from './AccordionSection';
-
-export { Accordion, AccordionSection, AccordionSectionHeader };
