--- ui/src/components/AccordionPanel/index.js
@@ -1,2 +1,0 @@
-import AccordionPanel from './AccordionPanel';
-export default AccordionPanel;
